kprobe: NAPI: trace __raise_softirq_irqoff_ksoft
Track packet handling through the NAPI execution path via a
kprobe that dumps a backtrace and prints out mode information the
first time it is hit, and then prints the number of hits when the
module is unloaded.  Takes the processor ID of the core where the
Eth and CAN IRQ threads are pinned as load-time parameters.

The code is based entirely on samples/kprobes/kprobe_example.c
and is based on a suggestion by Sebastian Siewor.
Example code for Oct, 2016 ELCE talk: http://sched.co/7rrr

Signed-off-by: Alison Chaiken <alison@she-devel.com>