kprobe: monitor NAPI by counting callers of do_current_softirqs
In the RT kernel, do_current_softirqs() may be invoked two
ways. The first way is when a hard IRQ handler exits and calls
__local_bh_enable().  The second is when ksoftirqd runs all the
softirqs that have not been raised by a hard IRQ (e.g. tasklets),
or have been raised when hard IRQs are disabled, which is true in
the case of NAPI processing.  This kprobe counts the occurrences
of the two execution paths for the NET_RX_SOFTIRQ on the
processor on which the network IRQ is running in order to
determine what fraction of packet handling occurs via NAPI and
what via hard IRQ handling.

Signed-off-by: Alison Chaiken <alison@she-devel.com>